Update API reference docs so that all links work
