+++ docs/reference/index.rst
@@ -2,6 +2,7 @@ API Reference
 =============
 
 .. toctree::
+    :maxdepth: 3
     :glob:
 
     py21cmmc*
