--- docs/usage.rst
@@ -1,10 +1,0 @@
-=====
-Usage
-=====
-
-On this page we provide a broad overview of the capabilities of ``21CMMC``, its
-philosophy, and how to use it. For more details, see either the :doc:`tutorials <tutorials>`
-or the :doc:`API Reference <reference/index>`, which itself contains several
-examples.
-
-.. note:: This page needs to be written!
