+++ src/py21cmmc/_utils.py
@@ -1,0 +1,12 @@
+"""Utility functions used throughout 21CMMC."""
+
+from collections import Iterable
+
+
+def flatten(items):
+    """Yield items from any nested iterable; see Reference."""
+    for x in items:
+        if isinstance(x, Iterable) and not isinstance(x, (str, bytes)):
+            yield from flatten(x)
+        else:
+            yield x
