+++ src/py21cmmc/analyse.py
@@ -10,7 +10,7 @@ import numpy as np
 from py21cmfast import yaml
 
 from .cosmoHammer import CosmoHammerSampler
-from .cosmoHammer.storage import HDFStorage
+from .cosmoHammer import HDFStorage
 
 
 def get_samples(chain, indx=0, burnin=False):
@@ -19,20 +19,21 @@ def get_samples(chain, indx=0, burnin=False):
 
     Parameters
     ----------
-    chain : :class:`~py21cmmc.mcmc.cosmoHammer.CosmoHammerSampler` or str
-        Either a `LikelihoodComputationChain`, which is the output of the :func:`~mcmc.run_mcmc` function,
-        or a path to an output HDF file containing the chain.
+    chain : :class:`~py21cmmc.cosmoHammer.CosmoHammerSampler` or str
+        Either a :class:`~py21cmmc.cosmoHammer.CosmoHammerSampler`, which is the output
+        of the :func:`~py21cmmc.mcmc.run_mcmc` function,
+        or a path to an output HDF5 file containing the chain.
 
     indx : int, optional
-        This is used only if `chain` is a string. It gives the index of the samples in the HDF file. Usually this is
-        zero.
+        This is used only if `chain` is a string. It gives the index of the samples in
+        the HDF file. Usually this is zero.
 
     burnin : bool
         Whether to return the burnin samples, rather than the actual run samples.
 
     Returns
     -------
-    store : a `HDFStore` object.
+    store : a :class:`~py21cmmc.cosmoHammer.HDFStore` object.
     """
     if isinstance(chain, CosmoHammerSampler):
         return (
@@ -54,16 +55,18 @@ def get_samples(chain, indx=0, burnin=False):
 
 def load_primitive_chain(modelname, direc="."):
     """
-    Load a chain produced by ``run_mcmc`` to be interactively useable.
+    Load a chain file produced by :func:`~py21cmmc.mcmc.run_mcmc` to be interactively useable.
 
     Parameters
     ----------
-    modelname : model name of the MCMC run.
-    direc : directory in which data was stored.
+    modelname : str
+        Model name of the MCMC run.
+    direc : str
+        Directory in which data was stored.
 
     Returns
     -------
-    chain : :class:`~py21cmmc.mcmc.cosmoHammer.LikelihoodComputationChain.LikelihoodComputationChain`
+    chain : :class:`~py21cmmc.cosmoHammer.LikelihoodComputationChain`
         The fully set-up chain, with no computed samples.
     """
     with open(join(direc, modelname + ".LCC.yml")) as f:
@@ -81,9 +84,9 @@ def corner_plot(
 
     Parameters
     ----------
-    samples: :class:`py21cmmc.mcmc.cosmoHammer.storage.HDFStorage`
-        The ``samples`` attribute of a sampler (i.e. the return value of :func:`~.mcmc.run_mcmc`), or equivalently,
-        the return value of :func:`~get_samples`.
+    samples: :class:`py21cmmc.cosmoHammer.HDFStorage` instance
+        The ``samples`` attribute of a sampler (i.e. the return value of
+        :func:`~.mcmc.run_mcmc`), or equivalently, the return value of :func:`~get_samples`.
     include_lnl: bool, optional
         Whether to plot the log-likelihood as if it were a parameter.
     show_guess: bool, optional
@@ -147,8 +150,9 @@ def trace_plot(
 
     Parameters
     ----------
-    samples: :class:`py21cmmc.mcmc.cosmoHammer.storage.HDFStorage`
-        The ``samples`` attribute of a sampler (i.e. the return value of :func:`~.mcmc.run_mcmc`), or equivalently,
+    samples: :class:`py21cmmc.cosmoHammer.HDFStorage`
+        The ``samples`` attribute of a sampler (i.e. the return value of
+        :func:`~py21cmmc.mcmc.run_mcmc`), or equivalently,
         the return value of :func:`~get_samples`.
     include_lnl: bool, optional
         Whether to plot the log-likelihood as if it were a parameter.
