--- src/py21cmmc/cosmoHammer/__init__.py
@@ -1,6 +1,0 @@
-"""A set of extensions to the basic CosmoHammer package."""
-
-from .CosmoHammerSampler import CosmoHammerSampler
-from .LikelihoodComputationChain import LikelihoodComputationChain
-from .storage import HDFStorageUtil
-from .util import Params
